5.AI_agent/requirements.txt
langchain
langchain_experimental
langchain-openai
tabulate
crewai-tools
python-dotenv
pandas